Update postcodes.csv
Yesterday I edited the EC06 error, which is a communication failure between the processor and RAM memories. My name is Duan. Today I realized that the error I edited was for codes 9134 and 9184. I don't know if it was an error when editing what I wrote or if the EC06 code is a subcode of 9184 and 9134. The ZQ test is a test I developed using an oscilloscope and measuring the pulses of the ZQ resistor of memories 1 per block A B C D, which can determine whether a block is operational or not. In the two links below, there is a video I made explaining it.

https://www.youtube.com/watch?v=hduecVL2Adc&t=310s

https://www.youtube.com/watch?v=ST2p3y9H4lI

In this second video, at minute 46 onwards, I test it.